Add initial support for NUMA [WIP]
Run make qemu-numa[-gdb] to emulate a NUMA RISC-V platform with CPUS
CPUs and 2 NUMA nodes.
For now, xv6 does not reach user space. Misconfiguration of the
VIRTIO_MMIO is suspected.

Mainly, the support is about modifying address computations for
socket-specific controllers and memory areas. In doing so, a few macros
have been renamed to more closely match the specification of the RISC-V
platform emulated by QEMU, which is the legacy SiFive.